fix: chat jobs not loading
